refactor(.): final refactor
